Refine the script to generate release package from release tag (#286)
* Update to support old tags build

* Remove Jenkins file and unused python script
